move all content ids to content management plugin
