+++ src/plugins/content_management/public/index.ts
@@ -14,3 +14,4 @@ export const plugin = (initializerContext: PluginInitializerContext) =>
 export * from './components';
 export * from './mocks';
 export * from './services/content_management';
+export * from './constants';
