+++ src/plugins/home/public/application/components/home_app.js
@@ -40,7 +40,8 @@ import { getTutorial } from '../load_tutorials';
 import { replaceTemplateStrings } from './tutorial/replace_template_strings';
 import { getServices } from '../opensearch_dashboards_services';
 import { useMount } from 'react-use';
-import { USE_NEW_HOME_PAGE, HOME_PAGE_ID } from '../../../common/constants';
+import { USE_NEW_HOME_PAGE } from '../../../common/constants';
+import { HOME_PAGE_ID } from '../../../../content_management/public';
 
 const RedirectToDefaultApp = () => {
   useMount(() => {
