+++ src/plugins/saved_objects_management/public/plugin.ts
@@ -65,7 +65,11 @@ import { registerServices } from './register_services';
 import { bootstrap } from './ui_actions_bootstrap';
 import { DEFAULT_NAV_GROUPS } from '../../../core/public';
 import { RecentWork } from './management_section/recent_work';
-import { HOME_CONTENT_AREAS } from '../../../plugins/home/public';
+import {
+  HOME_CONTENT_AREAS,
+  ESSENTIAL_OVERVIEW_CONTENT_AREAS,
+  ANALYTICS_ALL_OVERVIEW_CONTENT_AREAS,
+} from '../../../plugins/content_management/public';
 import { getScopedBreadcrumbs } from '../../opensearch_dashboards_react/public';
 import { NavigationPublicPluginStart } from '../../../plugins/navigation/public';
 
@@ -238,8 +242,8 @@ export class SavedObjectsManagementPlugin
       },
       getTargetArea: () => [
         HOME_CONTENT_AREAS.RECENTLY_VIEWED,
-        'analytics_overview/recently_viewed',
-        'all_overview/recently_viewed',
+        ESSENTIAL_OVERVIEW_CONTENT_AREAS.RECENTLY_VIEWED,
+        ANALYTICS_ALL_OVERVIEW_CONTENT_AREAS.RECENTLY_VIEWED,
       ],
     });
