+++ src/plugins/workspace/common/constants.ts
@@ -150,23 +150,3 @@ export const CURRENT_USER_PLACEHOLDER = '%me%';
 
 export const MAX_WORKSPACE_NAME_LENGTH = 40;
 export const MAX_WORKSPACE_DESCRIPTION_LENGTH = 200;
-// pattern for the page id is useCaseId_overview
-export const ESSENTIAL_OVERVIEW_PAGE_ID = 'analytics_overview';
-export const ANALYTICS_ALL_OVERVIEW_PAGE_ID = 'all_overview';
-export enum SECTIONS {
-  GET_STARTED = `get_started`,
-  SERVICE_CARDS = `service_cards`,
-  RECENTLY_VIEWED = `recently_viewed`,
-}
-
-export enum ESSENTIAL_OVERVIEW_CONTENT_AREAS {
-  GET_STARTED = `${ESSENTIAL_OVERVIEW_PAGE_ID}/${SECTIONS.GET_STARTED}`,
-  SERVICE_CARDS = `${ESSENTIAL_OVERVIEW_PAGE_ID}/${SECTIONS.SERVICE_CARDS}`,
-  RECENTLY_VIEWED = `${ESSENTIAL_OVERVIEW_PAGE_ID}/${SECTIONS.RECENTLY_VIEWED}`,
-}
-
-export enum ANALYTICS_ALL_OVERVIEW_CONTENT_AREAS {
-  GET_STARTED = `${ANALYTICS_ALL_OVERVIEW_PAGE_ID}/${SECTIONS.GET_STARTED}`,
-  SERVICE_CARDS = `${ANALYTICS_ALL_OVERVIEW_PAGE_ID}/${SECTIONS.SERVICE_CARDS}`,
-  RECENTLY_VIEWED = `${ANALYTICS_ALL_OVERVIEW_PAGE_ID}/${SECTIONS.RECENTLY_VIEWED}`,
-}
