+++ src/plugins/workspace/opensearch_dashboards.json
@@ -9,5 +9,5 @@
     "navigation"
   ],
   "optionalPlugins": ["savedObjectsManagement","management","dataSourceManagement","contentManagement"],
-  "requiredBundles": ["opensearchDashboardsReact", "home","dataSource"]
+  "requiredBundles": ["opensearchDashboardsReact","dataSource", "contentManagement"]
 }
